# app/migration-backup/api-zod/src/generated/types/quote.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteInput } from './quoteInput';
import type { QuoteStatus } from './quoteStatus';

export type Quote = QuoteInput & {
  id: string;
  createdAt: Date;
  status: QuoteStatus;
};