# app/migration-backup/api-zod/src/generated/types/quotePricing.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface QuotePricing {
  /** @minimum 0 */
  foodPerPlate: number;
  /** @minimum 0 */
  subtotal: number;
  gstEnabled: boolean;
  /**
     * @minimum 0
     * @maximum 28
     */
  gstRate: number;
  /** @minimum 0 */
  gstAmount: number;
  /** @maxLength 15 */
  gstNumber: string;
  gstNumberVisible: boolean;
  /** @minimum 0 */
  grandTotal: number;
  /** @minimum 0 */
  costPerGuest: number;
}