# app/migration-backup/api-zod/src/generated/types/quoteService.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { QuoteServicePricingType } from './quoteServicePricingType';

export interface QuoteService {
  id: string;
  name: string;
  category: string;
  subcategory: string;
  pricingType: QuoteServicePricingType;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  unitPrice: number;
  /** @minimum 0 */
  total: number;
}