# app/migration-backup/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: catalog
    description: Event service catalog operations
  - name: quotes
    description: Customer estimate operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /catalog:
    get:
      operationId: getCatalog
      tags: [catalog]
      summary: Get the active service catalog
      responses:
        "200":
          description: Active catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogCategory"
    put:
      operationId: updateCatalog
      tags: [catalog]
      summary: Replace the active service catalog
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/CatalogCategory"
      responses:
        "200":
          description: Updated active catalog
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CatalogCategory"
        "400":
          description: Invalid catalog
  /quotes:
    get:
      operationId: listQuotes
      tags: [quotes]
      summary: List recent estimates
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 25
      responses:
        "200":
          description: Recent estimates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Quote"
    post:
      operationId: createQuote
      tags: [quotes]
      summary: Save an estimate for follow-up
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteInput"
      responses:
        "201":
          description: Saved estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "400":
          description: Invalid estimate
  /quotes/{reference}:
    get:
      operationId: getQuote
      tags: [quotes]
      summary: Get an estimate by reference
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Estimate
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quote"
        "404":
          description: Estimate not found
    delete:
      operationId: deleteQuote
      tags: [quotes]
      summary: Delete a saved estimate
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Estimate deleted
        "404":
          description: Estimate not found
components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    CatalogItem:
      type: object
      required:
        - id
        - name
        - description
        - categoryId
        - subcategoryId
        - price
        - pricingType
        - active
        - minQuantity
        - maxQuantity
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        categoryId: { type: string }
        subcategoryId: { type: string }
        price: { type: number }
        pricingType:
          type: string
          enum: [perPerson, fixed, perUnit]
        active: { type: boolean }
        minQuantity: { type: integer, minimum: 1 }
        maxQuantity: { type: integer, minimum: 1 }
    CatalogSubcategory:
      type: object
      required: [id, name, items]
      properties:
        id: { type: string }
        name: { type: string }
        items:
          type: array
          items:
            $ref: "#/components/schemas/CatalogItem"
    CatalogCategory:
      type: object
      required: [id, name, description, subcategories]
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        subcategories:
          type: array
          items:
            $ref: "#/components/schemas/CatalogSubcategory"
    QuoteCustomer:
      type: object
      required: [name, mobile, eventType, eventDate, guests, hall]
      properties:
        name: { type: string, minLength: 1 }
        mobile: { type: string, minLength: 10 }
        eventType: { type: string, minLength: 1 }
        eventDate: { type: string, format: date }
        guests: { type: integer, minimum: 1 }
        gstNumber: { type: string, maxLength: 15 }
        hall: { type: string, minLength: 1 }
    QuoteService:
      type: object
      required: [id, name, category, subcategory, pricingType, quantity, unitPrice, total]
      properties:
        id: { type: string }
        name: { type: string }
        category: { type: string }
        subcategory: { type: string }
        pricingType:
          type: string
          enum: [perPerson, fixed, perUnit]
        quantity: { type: integer, minimum: 1 }
        unitPrice: { type: number, minimum: 0 }
        total: { type: number, minimum: 0 }
    QuotePricing:
      type: object
      required: [foodPerPlate, subtotal, gstEnabled, gstRate, gstAmount, gstNumber, gstNumberVisible, grandTotal, costPerGuest]
      properties:
        foodPerPlate: { type: number, minimum: 0 }
        subtotal: { type: number, minimum: 0 }
        gstEnabled: { type: boolean }
        gstRate: { type: number, minimum: 0, maximum: 28 }
        gstAmount: { type: number, minimum: 0 }
        gstNumber: { type: string, maxLength: 15 }
        gstNumberVisible: { type: boolean }
        grandTotal: { type: number, minimum: 0 }
        costPerGuest: { type: number, minimum: 0 }
    QuoteInput:
      type: object
      required: [reference, customer, services, pricing]
      properties:
        reference: { type: string, minLength: 1 }
        customer:
          $ref: "#/components/schemas/QuoteCustomer"
        services:
          type: array
          items:
            $ref: "#/components/schemas/QuoteService"
        pricing:
          $ref: "#/components/schemas/QuotePricing"
    Quote:
      allOf:
        - $ref: "#/components/schemas/QuoteInput"
        - type: object
          required: [id, createdAt, status]
          properties:
            id: { type: string }
            createdAt: { type: string, format: date-time }
            status:
              type: string
              enum: [new, contacted, confirmed, archived]